[syft/action_service] soft deleting an action ojbect / twin object from store

+++ packages/syft/src/syft/service/action/action_service.py
@@ -861,7 +861,7 @@ class ActionService(AbstractService):
 
     @service_method(path="action.delete", name="delete", roles=ADMIN_ROLE_LEVEL)
     def delete(
-        self, context: AuthedServiceContext, uid: UID
+        self, context: AuthedServiceContext, uid: UID, soft_delete: bool = True
     ) -> SyftSuccess | SyftError:
         get_res = self.store.get(uid=uid, credentials=context.credentials)
         if get_res.is_err():
@@ -876,7 +876,9 @@ class ActionService(AbstractService):
         return_msg.append(blob_del_res.message)
 
         # delete the action object from the action store
-        store_del_res = self._delete_from_action_store(context=context, obj=obj)
+        store_del_res = self._delete_from_action_store(
+            context=context, uid=obj.id, soft_delete=soft_delete
+        )
         if isinstance(store_del_res, SyftError):
             return SyftError(message=store_del_res.message)
         return_msg.append(store_del_res.message)
@@ -923,13 +925,45 @@ class ActionService(AbstractService):
     def _delete_from_action_store(
         self,
         context: AuthedServiceContext,
-        obj: TwinObject | ActionObject,
+        uid: UID,
+        soft_delete: bool = True,
     ) -> SyftSuccess | SyftError:
-        uid = obj.id
-        res = self.store.delete(credentials=context.credentials, uid=uid)
-        if res.is_err():
-            return SyftError(message=res.err())
+        if soft_delete:
+            get_res = self.store.get(uid=uid, credentials=context.credentials)
+            if get_res.is_err():
+                return SyftError(message=get_res.err())
+            obj: ActionObject | TwinObject = get_res.ok()
+            if isinstance(obj, TwinObject):
+                res = self._soft_delete(context=context, action_obj=obj.private)
+                if res.is_err():
+                    return SyftError(message=res.err())
+                res = self._soft_delete(context=context, action_obj=obj.mock)
+                if res.is_err():
+                    return SyftError(message=res.err())
+            else:  # ActionObject
+                res = self._soft_delete(context, obj)
+                if res.is_err():
+                    return SyftError(message=res.err())
+        else:
+            res = self.store.delete(credentials=context.credentials, uid=uid)
+            if res.is_err():
+                return SyftError(message=res.err())
+
         return SyftSuccess(message=f"Action object with uid '{uid}' deleted.")
+
+    def _soft_delete(
+        self, context: AuthedServiceContext, action_obj: ActionObject
+    ) -> Result[ActionObject, str]:
+        action_obj.syft_blob_storage_entry_id = None
+        action_obj.syft_action_data_type = None
+        action_obj.syft_action_data_repr_ = None
+        action_obj.syft_action_data_str_ = None
+        action_obj.syft_action_data_server_id = None
+        action_obj.syft_action_data_cache = None
+        res = self.set(context=context, action_object=action_obj)
+        if isinstance(res, SyftError):
+            return Err(res.message)
+        return Ok(action_obj)
 
 
 def resolve_action_args(
